requirements.txt
feedparser==6.0.11
requests==2.32.3
jinja2==3.1.4
python-dateutil==2.9.0
pytz==2024.2
xhtml2pdf==0.2.16
openpyxl==3.1.5
